[Sketches] factor out temperature sensor driver
Test with both IDEs so that command and demo run on both.

On Ubuntu 12.04 Linux it is necessary to hold the reset on the
LaunchPad while plugging in the USB and keep the reset depressed until
the serial monitor (or other terminal application) is loaded before
releasing the reset or the /dev/ttyACM0 just returns errors.
(probably a bug with ttyACM driver)

The demo on Arduino will only start cycling through the images when
the serial monitor is opened since there is a wait in the code for
proper USB connection on the Leonardo.

Signed-off-by: Christopher Hall <christopherhall.hsw@gmail.com>