[thermo] add note mentioning change to EPD_GFX.h to match display size
Problem is that EPD_GFX.h allocates a pair of fixed buffers
corresponding to the display size in the class.  This size is
determined at compile time.

Current solution is to manually modify the file to keep synchronised
with thermo.ino.  This has the advantage of not wasting memory when
the smaller display is used, and does not require a dynamic memory
allocator which can be problematic on low memory systems.
Disadvantage is having to modify two files.

Signed-off-by: Christopher Hall <christopherhall.hsw@gmail.com>