[ExampleScripts] Arch install updates and systemd configuration for epd_fuse driver
This explains how to get epd_fuse running on Arch Linux for Raspberry Pi
and Beaglebone Black.  Also the Arch fonts are on a different path.

Not fixed:  "import Image"
need to change to: from PIL import Image
there is a sed command in the README-Arch.md
as a temporary work around

Signed-off-by: Christopher Hall <christopherhall.hsw@gmail.com>