[Sketches] update demo for new driver, add script to create examples
The Ruby script make-examples will read selected sketches,
edit the file substituting:

  {% SECTION:item %}

with an appropriate value and write the sketch to:

  libraries/EPD_Vxxx_Gy/examples/NAME/NAME.ino

So that that a separate example file foe each possible combination of
example program, driver and panel size is created.  This allows for
just select and download from Arduino IDE - Noe editing the .ino is
required in this case.

Signed-off-by: Christopher Hall <christopherhall.hsw@gmail.com>