Avoid crash when spi not yet available
Under certain circumstances the spi device (/dev/spidev0.0) is not yet
available when epd-fuse starts. This resulted in not initialising the
epd structure and hence in a crash of epd-fuse at the first use.
Solved by letting udev signalling to systemd when spidev0.0 is made
available (file 98-spidev.rules in /etc/udev/rules.d) and modifying the
epd-fuse.service file to wait for /dev/spidev0.0