Fixes for Debian 11 Bullseye on RPi
In Bullseye the bcm_host.h and libbcm_host.a are moved to the standard include and library directories. In Buster and older versions they were in a special directory tree /opt/vc.
Moving LDFLAGS to the end of the comman line makes it work on both Bullseye and Buster